communication/requirements.txt
redis
colorama
toml
pyzmq